Move candidate_sampler to scripts
